suppress some compiler warnings
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@1168 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ src/texk/web2c/luatexdir/tex/texnodes.c
@@ -1213,7 +1213,7 @@ void print_free_chain(int c)
     halfword p = free_chain[c];
     fprintf(stdout, "\nfree chain[%d] =\n  ", c);
     while (p != null) {
-        fprintf(stdout, "%d,", p);
+	  fprintf(stdout, "%d,", (int)p);
         p = vlink(p);
     }
     fprintf(stdout, "null;\n");
